Move audio driver initializatio to thread pool More profiling scope added Fixed Canvas render target resizing problem Added TextureLoaderTest::stressTest

+++ roar/render/roCanvas.h
@@ -133,8 +133,8 @@ struct Canvas
 	roRDriverShaderTextureInput _textureInput;
 	StaticArray<roRDriverShaderBufferInput, 4> _bufferInputs;
 
-	float _targetWidth;
-	float _targetHeight;
+	unsigned _targetWidth;
+	unsigned _targetHeight;
 
 	// For image draw batching
 	bool _isBatchMode;
